choice for historical or real time dataviews
