+++ apps/fishing-map/features/workspace/Workspace.tsx
@@ -1,24 +1,15 @@
-import { useCallback, useEffect, useState } from 'react'
+import { Fragment, useCallback, useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
 import { useSelector } from 'react-redux'
 import { DndContext } from '@dnd-kit/core'
 import { restrictToVerticalAxis } from '@dnd-kit/modifiers'
 import { arrayMove } from '@dnd-kit/sortable'
 
-import { useLocalStorage } from '@globalfishingwatch/react-hooks'
-import { Button, IconButton, InputText, Modal, Spinner } from '@globalfishingwatch/ui-components'
+import type { ChoiceOption } from '@globalfishingwatch/ui-components'
+import { Choice, Spinner } from '@globalfishingwatch/ui-components'
 
-import { PUBLIC_SUFIX, ROOT_DOM_ELEMENT, USER_SUFIX } from 'data/config'
-import { WIZARD_TEMPLATE_ID } from 'data/highlighted-workspaces/marine-manager.dataviews'
-import {
-  DEEP_SEA_MINING_WORKSPACE_ID,
-  DEFAULT_WORKSPACE_ID,
-  WorkspaceCategory,
-} from 'data/workspaces'
 import { useAppDispatch } from 'features/app/app.hooks'
-import { selectReadOnly } from 'features/app/selectors/app.selectors'
 import { selectDataviewInstancesMergedOrdered } from 'features/dataviews/selectors/dataviews.resolvers.selectors'
-import { getModalParent } from 'features/modals/modals.utils'
 import { useFetchDataviewResources } from 'features/resources/resources.hooks'
 import { useUserExpiredToast } from 'features/user/user-expired.hooks'
 import { selectWorkspaceVessselGroupsIds } from 'features/vessel-groups/vessel-groups.selectors'
@@ -26,26 +17,23 @@ import {
   fetchWorkspaceVesselGroupsThunk,
   selectWorkspaceVesselGroupsStatus,
 } from 'features/vessel-groups/vessel-groups.slice'
-import type { WelcomeLocalStorageKey } from 'features/welcome/Welcome'
-import { DEEP_SEA_MINING_POPUP } from 'features/welcome/Welcome'
 import DetectionsSection from 'features/workspace/detections/DetectionsSection'
 import { useHideLegacyActivityCategoryDataviews } from 'features/workspace/legacy-activity-category.hook'
 import UserSection from 'features/workspace/user/UserSection/UserSection'
 import {
   selectIsWorkspacePasswordRequired,
+  selectTimeMode,
   selectWorkspace,
   selectWorkspaceStatus,
 } from 'features/workspace/workspace.selectors'
-import { setWorkspaceProperty } from 'features/workspace/workspace.slice'
-import { getWorkspaceLabel, isPrivateWorkspaceNotAllowed } from 'features/workspace/workspace.utils'
+import { isPrivateWorkspaceNotAllowed } from 'features/workspace/workspace.utils'
 import { useMigrateWorkspaceToast } from 'features/workspace/workspace-migration.hooks'
 import WorkspaceError from 'features/workspace/WorkspaceError'
 import WorkspacePassword from 'features/workspace/WorkspacePassword'
-import { updateWorkspaceThunk } from 'features/workspaces-list/workspaces-list.slice'
+import WorkspaceTitle from 'features/workspace/WorkspaceTitle'
 import { useReplaceQueryParams } from 'router/routes.hook'
-import { selectLocationCategory } from 'router/routes.selectors'
+import type { TimeMode } from 'types'
 import { AsyncReducerStatus } from 'utils/async-slice'
-import { htmlSafeParse } from 'utils/html-parser'
 
 import ActivitySection from './activity/ActivitySection'
 import ContextAreaSection from './context-areas/ContextAreaSection'
@@ -57,38 +45,19 @@ import VesselsSection from './vessels/VesselsSection'
 import styles from './Workspace.module.css'
 
 function Workspace() {
+  const { t } = useTranslation()
   useHideLegacyActivityCategoryDataviews()
   useUserExpiredToast()
   useMigrateWorkspaceToast()
-  const { t } = useTranslation()
   const dispatch = useAppDispatch()
   const { replaceQueryParams } = useReplaceQueryParams()
-  const readOnly = useSelector(selectReadOnly)
   const isWorkspacePasswordRequired = useSelector(selectIsWorkspacePasswordRequired)
   const workspace = useSelector(selectWorkspace)
+  const timeMode = useSelector(selectTimeMode)
   const dataviews = useSelector(selectDataviewInstancesMergedOrdered)
   const workspaceStatus = useSelector(selectWorkspaceStatus)
   const workspaceVesselGroupsStatus = useSelector(selectWorkspaceVesselGroupsStatus)
-  const locationCategory = useSelector(selectLocationCategory)
   const workspaceVesselGroupsIds = useSelector(selectWorkspaceVessselGroupsIds)
-  const isUserWorkspace =
-    workspace?.id?.endsWith(`-${USER_SUFIX}`) ||
-    workspace?.id?.endsWith(`-${USER_SUFIX}-${PUBLIC_SUFIX}`)
-  const [workspaceEditName, setWorkspaceEditName] = useState(workspace?.name)
-  const [workspaceEditDescription, setWorkspaceEditDescription] = useState(workspace?.description)
-  const [workspaceEditModalOpen, setWorkspaceEditModalOpen] = useState(false)
-  const [editWorkspaceLoading, setEditWorkspaceLoading] = useState(false)
-  const [dsmWelcomePopup, setDsmWelcomePopup] = useLocalStorage<WelcomeLocalStorageKey>(
-    DEEP_SEA_MINING_POPUP,
-    { visible: false, showAgain: false }
-  )
-
-  useEffect(() => {
-    if (workspace) {
-      setWorkspaceEditName(workspace.name)
-      setWorkspaceEditDescription(workspace.description)
-    }
-  }, [workspace])
 
   useFetchDataviewResources()
 
@@ -111,29 +80,6 @@ function Workspace() {
       }
     },
     [dataviews, replaceQueryParams]
-  )
-
-  const onWorkspaceUpdateClose = useCallback(() => {
-    setEditWorkspaceLoading(false)
-    setWorkspaceEditModalOpen(false)
-  }, [])
-
-  const onWorkspaceUpdateClick = useCallback(
-    async (workspaceId: string) => {
-      setEditWorkspaceLoading(true)
-      await dispatch(
-        updateWorkspaceThunk({
-          id: workspaceId,
-          name: workspaceEditName,
-          description: workspaceEditDescription,
-        })
-      )
-      if (workspaceEditName) {
-        dispatch(setWorkspaceProperty({ key: 'name', value: workspaceEditName }))
-      }
-      onWorkspaceUpdateClose()
-    },
-    [dispatch, onWorkspaceUpdateClose, workspaceEditDescription, workspaceEditName]
   )
 
   if (
@@ -159,86 +105,38 @@ function Workspace() {
     return <WorkspaceError />
   }
 
-  const openDSMPopup = () => {
-    setDsmWelcomePopup({ ...dsmWelcomePopup, visible: true })
-  }
+  const options = [
+    {
+      id: 'historical',
+      label: t((t) => t.common.historical),
+    },
+    {
+      id: 'realTime',
+      label: t((t) => t.common.realTime),
+    },
+  ] as ChoiceOption<TimeMode>[]
 
   return (
     <DndContext onDragEnd={handleDragEnd} modifiers={[restrictToVerticalAxis]}>
-      {(locationCategory === WorkspaceCategory.MarineManager ||
-        locationCategory === WorkspaceCategory.FishingActivity) &&
-        workspace?.name &&
-        workspace?.id !== WIZARD_TEMPLATE_ID &&
-        workspace?.id !== DEFAULT_WORKSPACE_ID &&
-        !readOnly && (
-          <div className={styles.header}>
-            {isUserWorkspace && (
-              <label className={styles.subTitle}>{t((t) => t.workspace.user)}</label>
-            )}
-            <h2 className={styles.title} data-test="user-workspace-title">
-              {getWorkspaceLabel(workspace)}
-              {/* {gfwUser && (
-                <IconButton
-                  className="print-hidden"
-                  size="small"
-                  icon="edit"
-                  onClick={() => setWorkspaceEditModalOpen(true)}
-                />
-              )} */}
-            </h2>
-            {workspace?.id === DEEP_SEA_MINING_WORKSPACE_ID && (
-              <h3 className={styles.subTitle}>
-                {htmlSafeParse(workspace.description)}
-                <IconButton
-                  className={styles.subTitleBtn}
-                  icon="info"
-                  size="tiny"
-                  onClick={openDSMPopup}
-                />
-              </h3>
-            )}
-            <Modal
-              appSelector={ROOT_DOM_ELEMENT}
-              title={t((t) => t.workspace.edit)}
-              isOpen={workspaceEditModalOpen}
-              contentClassName={styles.modalContainer}
-              onClose={onWorkspaceUpdateClose}
-              parentSelector={getModalParent}
-            >
-              <div className={styles.content}>
-                <InputText
-                  value={workspaceEditName}
-                  className={styles.input}
-                  label={t((t) => t.common.name)}
-                  onChange={(e) => setWorkspaceEditName(e.target.value)}
-                />
-                <InputText
-                  value={workspaceEditDescription}
-                  className={styles.input}
-                  label={t((t) => t.common.description)}
-                  onChange={(e) => setWorkspaceEditDescription(e.target.value)}
-                />
-              </div>
-              <div className={styles.modalFooter}>
-                <Button
-                  className={styles.saveBtn}
-                  loading={editWorkspaceLoading}
-                  onClick={() => onWorkspaceUpdateClick(workspace?.id)}
-                >
-                  {t((t) => t.common.update) as string}
-                </Button>
-              </div>
-            </Modal>
-          </div>
-        )}
-      <ActivitySection />
-      <DetectionsSection />
-      <EventsSection />
-      <VesselsSection />
-      <VesselGroupSection />
-      <EnvironmentalSection />
-      <ContextAreaSection />
-      <UserSection />
+      <WorkspaceTitle />
+      <div className={styles.header}>
+        <Choice
+          size="medium"
+          options={options}
+          activeOption={timeMode}
+          onSelect={(option) => replaceQueryParams({ timeMode: option.id })}
+        />
+      </div>
+      <Fragment>
+        <ActivitySection />
+        {timeMode === 'historical' && <DetectionsSection />}
+        {timeMode === 'historical' && <EventsSection />}
+        <VesselsSection />
+        {timeMode === 'historical' && <VesselGroupSection />}
+        {timeMode === 'historical' && <EnvironmentalSection />}
+        <ContextAreaSection />
+        <UserSection />
+      </Fragment>
     </DndContext>
   )
 }
