+++ apps/fishing-map/features/workspace/workspace.selectors.ts
@@ -218,6 +218,7 @@ export const selectWorkspaceVisibleEventsArray = createSelector(
 
 export const selectDaysFromLatest = selectWorkspaceStateProperty('daysFromLatest')
 export const selectCollapsedSections = selectWorkspaceStateProperty('collapsedSections')
+export const selectTimeMode = selectWorkspaceStateProperty('timeMode')
 
 export type WorkspaceFetchParams = { workspaceId: string; reportId?: string }
