+++ apps/fishing-map/public/locales/source/translations.json
@@ -233,6 +233,7 @@
     "hideAllHelpHints": "Dismiss all",
     "hideHelpHint": "Dismiss",
     "hints": "Need help? Look for these dots",
+    "historical": "Historical",
     "hour": "Hour",
     "hour_one": "Hour",
     "hour_many": "hours",
@@ -294,6 +295,7 @@
     "points_other": "points",
     "presence": "Vessel presence",
     "previously": "Previously",
+    "realTime": "Real time",
     "report": "Report",
     "reports": "Reports",
     "resetHelpHints": "Show all help hints again",
